Nuri/update: Consolidate video banner comp into homepage (#1913)
* consolidate video banner comp into homepage and centralize the video player

* remove videoplayer file in video folder
