+++ src/pages/blog/video-banner/_DBanner.js
@@ -1,8 +1,8 @@
 import React, { useEffect, useState } from 'react'
 import styled from 'styled-components'
 import PropTypes from 'prop-types'
+import VideoPlayer from '../_video-player'
 import VideoCarousel from './_VideoCarousel'
-import VideoPlayer from './_video-player'
 import { Flex, Container } from 'components/containers'
 import { Header } from 'components/elements'
 import device from 'themes/device'
